requirements.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
contextlib2==0.5.5
cryptography==1.9
defusedxml==0.5.0
django==1.11.5
django-allauth==0.32.0
django-appconf==1.0.2
django-bootstrap3==8.2.3
django-compressor==2.1.1
django-cors-headers==2.1.0
django-helusers==0.3.1
django-modeltranslation==0.12.1
django-mptt==0.8.7
django-multiselectfield==0.1.7
django-munigeo==0.2.0
django-npm==1.0.0
django-oauth-toolkit==1.0.0
git+git://github.com/City-of-Helsinki/django-oidc-provider.git@v0.5.0.post1+uc#egg=django-oidc-provider==0.5.0.post1+uc
django-parler==1.8
django-sass-processor==0.5.4
django-thesaurus==0.0.1
djangorestframework==3.6.3
future==0.16.0
idna==2.5
isodate==0.5.4
libsass==0.13.2
oauthlib==2.0.2
psycopg2==2.7.1
pycparser==2.17
pycryptodomex==3.4.6
pyjwkest==1.3.0
pyjwt==1.5.2
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
pyyaml==3.12
raven==6.1.0
rcssmin==1.0.6
rdflib==4.2.2
requests==2.18.4
requests-cache==0.4.13
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.10.0
urllib3==1.22
